blog/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-mysql==4.9.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
